Move label actions back to YAML with shared actions
Keep the label action data in data/github/label_actions.yaml, but split
the configuration into a central actions section, where each action and
its comment is defined once, and a repositories section which lists the
enabled labels per repository. This keeps shared texts centralized
while allowing per-repository opt-in.
