+++ data/github/label_actions.yaml
@@ -1,0 +1,129 @@
+# Label actions.
+# When a label listed for a repository under "repositories" is added to an
+# issue, the action with the same name defined under "actions" is performed:
+# the configured comment is posted and, if close is set, the issue is closed
+# with the given closeReason (not_planned or completed).
+#
+# In comments, {issue-author} is replaced with the username of the issue
+# author and {repository} with the full name (owner/repo) of the repository
+# the issue was reported in.
+actions:
+  assume-fixed:
+    comment: >
+      :wave: @{issue-author}, thanks for reporting an issue!
+
+
+      This issue is assumed to be fixed in the latest stable release. Please
+      reopen in case you can still reproduce the issue with the latest stable
+      release. You can find the latest stable release at
+      https://github.com/{repository}/releases/latest
+    close: true
+    closeReason: not_planned
+
+  core-issue:
+    comment: >
+      :wave: @{issue-author}, thanks for reporting an issue!
+
+
+      It looks like this issue is related to Home Assistant Core. Please check
+      the [Home Assistant Core](https://github.com/home-assistant/core/issues)
+      repository, the issue might have been reported already. Open a new issue
+      in that repository if you can't find a matching issue.
+    close: true
+    closeReason: not_planned
+
+  frontend-issue:
+    comment: >
+      :wave: @{issue-author}, thanks for reporting an issue!
+
+
+      It looks like this issue is related to Home Assistant Frontend. Please
+      check the [Home Assistant Frontend](https://github.com/home-assistant/frontend/issues)
+      repository, the issue might have been reported already. Open a new issue
+      in that repository if you can't find a matching issue.
+    close: true
+    closeReason: not_planned
+
+  os-issue:
+    comment: >
+      :wave: @{issue-author}, thanks for reporting an issue!
+
+
+      It looks like this issue is related to Home Assistant Operating System.
+      Please check the [Home Assistant Operating System](https://github.com/home-assistant/operating-system/issues)
+      repository, the issue might have been reported already. Open a new issue
+      in that repository if you can't find a matching issue.
+    close: true
+    closeReason: not_planned
+
+  supervisor-issue:
+    comment: >
+      :wave: @{issue-author}, thanks for reporting an issue!
+
+
+      It looks like this issue is related to Home Assistant Supervisor. Please
+      check the [Home Assistant Supervisor](https://github.com/home-assistant/supervisor/issues)
+      repository, the issue might have been reported already. Open a new issue
+      in that repository if you can't find a matching issue.
+    close: true
+    closeReason: not_planned
+
+  new-feature:
+    comment: >
+      :wave: @{issue-author}, thanks for your input!
+
+      We use this issue tracker to track issues of currently supported features.
+      Your request appears to request a new feature. We track potential new
+      features in the [Feature Request section of our Community Forum](https://community.home-assistant.io/c/feature-requests/13).
+      Please check if someone already requested a similar feature, or create
+      a new feature request in that forum. Thank you!
+    close: true
+    closeReason: not_planned
+
+  docker-corruption:
+    comment: >
+      :wave: @{issue-author}, thanks for reporting an issue!
+
+
+      Home Assistant uses Docker container images under the hood to run all
+      parts of the system (Home Assistant Core, add-ons and other components).
+      The symptoms described here (e.g. 0-byte or empty files, missing files
+      or directories, or corrupt binaries inside a container) indicate that
+      this Docker image storage got corrupted on your system. This is a local
+      problem, typically caused by a power loss or forced reset while the
+      system was writing to disk, or by failing storage. It is not a problem
+      in the published image or in Home Assistant itself.
+
+
+      **What to do:** Create a full backup and download it to another machine.
+      Then reinstall Home Assistant OS and restore the backup. This rewrites
+      all container images and reliably resolves the problem.
+
+
+      For those familiar with Docker, some background on why this is the only
+      reliable fix: Docker verifies image checksums only while pulling. Once
+      an image is written to disk, its integrity is never checked again, so
+      corruption can stay dormant until the affected container is recreated,
+      e.g. on a host reboot. Deleting and re-pulling the affected image does
+      not necessarily help either: image layers are shared between images and
+      reused across versions, and `docker pull` skips layers it believes are
+      already present locally, so the corrupted data survives a re-pull. A
+      dedicated feature to reset the Docker image storage completely is
+      tracked in
+      [#6555](https://github.com/home-assistant/supervisor/issues/6555).
+    close: true
+    closeReason: not_planned
+
+repositories:
+  home-assistant/operating-system:
+    - assume-fixed
+    - core-issue
+    - frontend-issue
+    - supervisor-issue
+    - new-feature
+
+  home-assistant/supervisor:
+    - core-issue
+    - frontend-issue
+    - os-issue
+    - docker-corruption
